test_requirements.txt
coveralls<3.3
flake8
flake8-docstrings
hacking>=1.0
m2r2
numpydoc
pytest-cov
pytest-ordering
sphinx
sphinx_rtd_theme
